WIP: Remove code duplication
