+++ dev/src/bulk-writer.ts
@@ -19,184 +19,20 @@ import {FieldPath, Firestore} from '.';
 import {DocumentReference} from './reference';
 import {Timestamp} from './timestamp';
 import {DocumentData, Precondition, SetOptions, UpdateData} from './types';
-import {Deferred} from './util';
-import {BatchWriteResult, WriteBatch, WriteResult} from './write-batch';
+import {
+  BatchState,
+  MAX_BATCH_SIZE,
+  UpdateBuilder,
+  WriteResult
+} from "./update-builder";
 
-/**
- * The maximum number of writes that can be in a single batch.
- */
-const MAX_BATCH_SIZE = 500;
 
-/**
- * Used to represent the state of batch.
- *
- * Writes can only be added while the batch is OPEN. For a batch to be sent,
- * the batch must be READY_TO_SEND. After a batch is sent, it is marked as SENT.
- */
-enum BatchState {
-  OPEN,
-  READY_TO_SEND,
-  SENT,
+class BulkCommitBatch extends UpdateBuilder<Promise<WriteResult>> {
+  wrapResult(res: Promise<WriteResult>): Promise<WriteResult> {
+    return res;
+  }
+  
 }
-
-/**
- * Used to represent a batch on the BatchQueue.
- *
- * @private
- */
-class BulkCommitBatch {
-  /**
-   * The state of the batch.
-   */
-  state = BatchState.OPEN;
-
-  // The set of document reference paths present in the WriteBatch.
-  readonly docPaths = new Set<string>();
-
-  // A deferred promise that is resolved after the batch has been sent, and a
-  // response is received.
-  private completedDeferred = new Deferred<void>();
-
-  // A map from each WriteBatch operation to its corresponding result.
-  private resultsMap = new Map<number, Deferred<BatchWriteResult>>();
-
-  constructor(
-    private readonly writeBatch: WriteBatch,
-    private readonly maxBatchSize: number
-  ) {}
-
-  /**
-   * The number of writes in this batch.
-   *
-   * @property
-   */
-  get opCount(): number {
-    return this.resultsMap.size;
-  }
-
-  /**
-   * Adds a `create` operation to the WriteBatch. Returns a promise that
-   * resolves with the result of the write.
-   */
-  create(
-    documentRef: DocumentReference,
-    data: DocumentData
-  ): Promise<WriteResult> {
-    this.writeBatch.create(documentRef, data);
-    return this.processOperation(documentRef);
-  }
-
-  /**
-   * Adds a `delete` operation to the WriteBatch. Returns a promise that
-   * resolves with the result of the delete.
-   */
-  delete(
-    documentRef: DocumentReference,
-    precondition?: Precondition
-  ): Promise<WriteResult> {
-    this.writeBatch.delete(documentRef, precondition);
-    return this.processOperation(documentRef);
-  }
-
-  /**
-   * Adds a `set` operation to the WriteBatch. Returns a promise that
-   * resolves with the result of the write.
-   */
-  set(
-    documentRef: DocumentReference,
-    data: DocumentData,
-    options?: SetOptions
-  ): Promise<WriteResult> {
-    this.writeBatch.set(documentRef, data, options);
-    return this.processOperation(documentRef);
-  }
-
-  /**
-   * Adds an `update` operation to the WriteBatch. Returns a promise that
-   * resolves with the result of the write.
-   */
-  update(
-    documentRef: DocumentReference,
-    dataOrField: UpdateData | string | FieldPath,
-    ...preconditionOrValues: Array<
-      {lastUpdateTime?: Timestamp} | unknown | string | FieldPath
-    >
-  ): Promise<WriteResult> {
-    this.writeBatch.update(documentRef, dataOrField, ...preconditionOrValues);
-    return this.processOperation(documentRef);
-  }
-
-  /**
-   * Helper to update data structures associated with the operation and
-   * return the result.
-   */
-  private processOperation(
-    documentRef: DocumentReference
-  ): Promise<WriteResult> {
-    assert(
-      !this.docPaths.has(documentRef.path),
-      'Batch should not contain writes to the same document'
-    );
-    assert(
-      this.state === BatchState.OPEN,
-      'Batch should be OPEN when adding writes'
-    );
-    this.docPaths.add(documentRef.path);
-    const deferred = new Deferred<BatchWriteResult>();
-    this.resultsMap.set(this.opCount, deferred);
-
-    if (this.opCount === this.maxBatchSize) {
-      this.state = BatchState.READY_TO_SEND;
-    }
-
-    return deferred.promise.then(result => {
-      if (result.writeTime) {
-        return new WriteResult(result.writeTime);
-      } else {
-        throw result.status;
-      }
-    });
-  }
-
-  /**
-   * Commits the batch and returns a promise that resolves with the result of
-   * all writes in this batch.
-   */
-  bulkCommit(): Promise<BatchWriteResult[]> {
-    assert(
-      this.state === BatchState.READY_TO_SEND,
-      'The batch should be marked as READY_TO_SEND before committing'
-    );
-    this.state = BatchState.SENT;
-    return this.writeBatch.bulkCommit();
-  }
-
-  /**
-   * Resolves the individual operations in the batch with the results.
-   */
-  processResults(results: BatchWriteResult[]): void {
-    results.map((result, index) => {
-      this.resultsMap.get(index)!.resolve(result);
-    });
-    this.completedDeferred.resolve();
-  }
-
-  /**
-   * Returns a promise that resolves when the batch has been sent, and a
-   * response is received.
-   */
-  awaitBulkCommit(): Promise<void> {
-    this.markReadyToSend();
-    return this.completedDeferred.promise;
-  }
-
-  markReadyToSend(): void {
-    if (this.state === BatchState.OPEN) {
-      this.state = BatchState.READY_TO_SEND;
-    }
-  }
-}
-
 /**
  * A Firestore BulkWriter than can be used to perform large amounts of writes in
  * parallel. Writes to the same document will be executed sequentially.
@@ -212,7 +48,7 @@ export class BulkWriter {
   /**
    * A queue of batches to be written.
    */
-  private batchQueue: BulkCommitBatch[] = [];
+  private batchQueue: UpdateBuilder<Promise<WriteResult>>[] = [];
 
   /**
    * Whether this BulkWriter instance is closed. Once closed, it cannot be
@@ -485,7 +321,7 @@ export class BulkWriter {
 
   private createNewBatch(): BulkCommitBatch {
     const newBatch = new BulkCommitBatch(
-      this.firestore.batch(),
+      this.firestore,
       this.maxBatchSize
     );
